Operabilità Rotte.xlsx is an Excel prototype
The prototype is in Italian as it is mainly directed to an Italian audience.
Operabilità Rotte.xlsx is an Excel prototype modelling the operability of airline routes.
In this version there are 2 sheets and 3 charts.
The first sheet allow the setting of the input parameters. The default settings are for a Ryanair flight from Reggio Calabria to Milan Linate.
The last sheet host the modeller.
The first chart shows the operability in term of required load factors. The default settings shows that the sample route, given the input parameters, is not operable at any level of load factor. Playing with the inputs may change this outcome, e.g. zeroing the government tax (imposte) will make the route operable.
The second chart shows actual costs vs actual revenue. On the x axis we see the incremental number of passengers buying a ticket, on the y axis we see the financial consequences in costs and revenues. 
The third chart shows the relationship between actual costs and actual profits. On the x axis we see the incremental number of passengers buying a ticket, on the y axis we see the financial consequences in costs and profits.